chatbot/requirements.txt
pymysql==1.0.3
python-dotenv==1.0.0
rasa==3.4.5
rasa_sdk==3.4.1
requests==2.28.2
websockets==10.4